+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Handling the case when the plugins info contains invalid data. #110
+- Do not install dev build plugins when looking for the latest version (unless --include-dev-builds is used). #109
 
 ### Added